sdm: Do not mandatory enable demura while set power on
SDM add demura dynamic enable and disable supports.
If demura was set to disabled, it should not enable
it after power on. But demura is mandatory to
disable/enable while set power off/on.
Change will not enable demura in power on if it has
been set to disabled dynamically.

Change-Id: I08cd58d038884b8f504ac675ed660aa3571320af