composer: add support for front buffer rendering
Add support for front buffer layers. HWC must not perform any
caching for this buffer when submitted for composition. HWCs
do not need to do any form of auto-refresh, and they are allowed
to cache composition results between presents from SF (such as
for panel self-refresh), but for any given present the buffer
must be composited from even if it otherwise appears to be the
same as a previous composition.

Change-Id: I84f17023702421321144ae934e0767d5b85b0805