hwc: Introduce property to enable content metadata dumping during frame dump
Introduce a property to allow content metadata dumping as part of
the HWC frame dump binder call. This property is intended to be used
for debugging and test development.

Only custom content metadata is currently dumped when this property
is enabled. It is intended to be extendable to dump other metadata in
the future.

Change-Id: Ibc017c0363853210d59ac5bae082faae13e44c21