hwc: Fix HPD race condition
Currently there is HWCUEvent class which creates a static object
and receives HWCUEventListener pointer from HWCSession.
As HWCUEvent is static, it will start first while its handler pointer
may remain null until HWCSession comes up.

Simplify this by starting the HPD threads in HWCSession itself. So there
is no need to provide handler as pointer.

The Top thread needs to be detached from main thread as it waits on
uevent_next_event() which can't be controlled from main thread.

The bottom thread can be attached to main thread. When HWCSession is
shutting down, we can wake up the bottom thread and wait for it to end.

Change-Id: I059f57373cc2c75438fbafa2ade65a9484860b48