sdm: Enable FB ID caching for Demura
A framebuffer object gets created for demura layers every draw cycle.
Due to the recent change where the plane properties are configured to
the driver whenever the FB is updated or renewed, HWC's CPU increases
when demura is enabled.

This change enables FB ID caching for demura to prevent SDM from
creating new FB ID for the layer in every frame.

//Recent change-Id: I52cdeb96585fa1a5e74050d2855898b3b9b2ebaf

Change-Id: I945e1c08307bfa42f799a98c627e40506a9b6c77
CRs-Fixed: 3560880