hwc: Use warning messages for connector lists
User-space receives multiple events from driver (including from upstream driver)
when DP MST is connected or disconnected. User-space process each event by
reading the connector list. When disconnect is handled by user-space, it cleans
up its data and ready to process next event. While it processes next event,
driver might still be cleaning up the connector state. In these race conditions,
it is highly likely that user-space would read a connector in invalid state as it
is currently being destroyed or modified in response to previous disconnect
request.

Such cases are not really an error, mark them as warning just to show that a
connector could not be fully retrieved. As the connector data has been fully
released in user-space, it is safe to ignore these errors.

Change-Id: Idd7a66360b9c2d492170b8a31a0d13ad8fff973c