Composer: fix HWC failed to be set affinity to gold cores when large Comp boost.
Composer sometimes could not send the initial hwcPid to perf hal, which
leads the hwc failed to be set affinity to gold cores when large
composition boost.

Add return value check for ReqHintsOffload, which can avoid the hwc_tid_
to be overwritten by tid. So the ReqTidChangeOffload function can be
called normally. And the initial hwcPid can be sent to perf hal
correctly to be set affinity to gold cores when large comp boost.

CRs-Fixed: 3775007
Change-Id: Ib7a8d9583eec8b9005f66be90c85bc78601e0d63